workers/requirements.txt
# Code formatter
black

# Debug
pydevd-pycharm==193.5233.109

# HTTP requests handler
requests

# Multipart handler
requests-toolbelt

# Flask
Flask
Flask-SQLAlchemy
Flask-SocketIO

# MySQL
mysqlclient

# SQLAlchemy
SQLAlchemy

# Keycloak (alternative client for token exchange)
python-keycloak-client

# Celery
Celery

# Redis
redis

# JSON
simplejson

# Serialization
jsonpickle

# Computation modules (should disappear)
pyradiomics

# YAML
pyyaml